+++ lib/components/achievements/achievement_dialog.dart
@@ -8,18 +8,18 @@ import 'package:flutter_svg/svg.dart';
 import 'package:intl/intl.dart';
 import 'package:share_plus/share_plus.dart';
 
-import '../flavors.dart';
-import '../graphql/queries/achievements.graphql.dart';
-import '../theme/design_system/design_system.dart';
-import '../helpers/ui/image.dart';
-import '../helpers/ui/svg_icons.dart';
-import '../l10n/app_localizations.dart';
-import '../providers/analytics.dart';
-import '../helpers/share_image.dart';
-import '../helpers/utils.dart';
-import '../helpers/ui/widget_to_image.dart';
-import 'dialog_with_image.dart';
-import 'loading_indicator.dart';
+import '../../flavors.dart';
+import '../../graphql/queries/achievements.graphql.dart';
+import '../../theme/design_system/design_system.dart';
+import '../../helpers/ui/image.dart';
+import '../../helpers/ui/svg_icons.dart';
+import '../../l10n/app_localizations.dart';
+import '../../providers/analytics.dart';
+import '../../helpers/share_image.dart';
+import '../../helpers/utils.dart';
+import '../../helpers/ui/widget_to_image.dart';
+import '../dialog_with_image.dart';
+import '../loading_indicator.dart';
 
 class AchievementDialog extends ConsumerStatefulWidget {
   const AchievementDialog({super.key, required this.achievement, this.dismissButtonText});
