+++ lib/components/achievements/achievement_list.dart
@@ -3,10 +3,10 @@ import 'package:brunstadtv_app/providers/analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../graphql/queries/achievements.graphql.dart';
+import '../../graphql/queries/achievements.graphql.dart';
 import 'achievement_dialog.dart';
 import 'achievement_list_item.dart';
-import 'custom_grid_view.dart';
+import '../custom_grid_view.dart';
 
 class AchievementList extends ConsumerWidget {
   const AchievementList({super.key, required this.achievements});
